+++ tests/src/lib.rs
@@ -1,6 +1,9 @@
 #![feature(stmt_expr_attributes)]
+mod env;
+mod file_ops;
+mod http;
 mod list_targets;
-mod sanity;
+mod traffic;
 
 use std::{
     collections::HashMap,
