+++ tests/src/traffic.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 
-mod tests {
+mod traffic {
     use std::{net::UdpSocket, time::Duration};
 
     use bytes::Bytes;
@@ -9,246 +9,8 @@ mod tests {
     use k8s_openapi::api::core::v1::Pod;
     use kube::{api::LogParams, Api, Client};
     use rstest::*;
-    use tokio::time::timeout;
 
     use crate::*;
-
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    #[rstest]
-    #[trace]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    async fn test_mirror_http_traffic(
-        #[future]
-        #[notrace]
-        service: KubeService,
-        #[future]
-        #[notrace]
-        kube_client: Client,
-        #[values(
-            Application::NodeHTTP,
-            Application::Go18HTTP,
-            Application::Go19HTTP,
-            Application::PythonFlaskHTTP,
-            Application::PythonFastApiHTTP
-        )]
-        application: Application,
-        #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
-    ) {
-        let service = service.await;
-        let kube_client = kube_client.await;
-        let url = get_service_url(kube_client.clone(), &service).await;
-        let mut process = application
-            .run(
-                &service.target,
-                Some(&service.namespace),
-                agent.flag(),
-                None,
-            )
-            .await;
-        process.wait_for_line(Duration::from_secs(120), "daemon subscribed");
-        send_requests(&url, false).await;
-        process.wait_for_line(Duration::from_secs(10), "GET");
-        process.wait_for_line(Duration::from_secs(10), "POST");
-        process.wait_for_line(Duration::from_secs(10), "PUT");
-        process.wait_for_line(Duration::from_secs(10), "DELETE");
-        timeout(Duration::from_secs(40), process.child.wait())
-            .await
-            .unwrap()
-            .unwrap();
-
-        application.assert(&process);
-    }
-
-    #[ignore] // TODO: create integration test instead.
-    #[cfg(target_os = "macos")]
-    #[rstest]
-    #[trace]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    async fn test_mirror_http_traffic(
-        #[future]
-        #[notrace]
-        service: KubeService,
-        #[future]
-        #[notrace]
-        kube_client: Client,
-        #[values(Application::PythonFlaskHTTP, Application::PythonFastApiHTTP)]
-        application: Application,
-        #[values(Agent::Job)] agent: Agent,
-    ) {
-        let service = service.await;
-        let kube_client = kube_client.await;
-        let url = get_service_url(kube_client.clone(), &service).await;
-        let mut process = application
-            .run(
-                &service.target,
-                Some(&service.namespace),
-                agent.flag(),
-                None,
-            )
-            .await;
-        process.wait_for_line(Duration::from_secs(300), "daemon subscribed");
-        send_requests(&url, false).await;
-        process.wait_for_line(Duration::from_secs(10), "GET");
-        process.wait_for_line(Duration::from_secs(10), "POST");
-        process.wait_for_line(Duration::from_secs(10), "PUT");
-        process.wait_for_line(Duration::from_secs(10), "DELETE");
-        timeout(Duration::from_secs(40), process.child.wait())
-            .await
-            .unwrap()
-            .unwrap();
-
-        application.assert(&process);
-    }
-
-    #[cfg(target_os = "linux")]
-    #[rstest]
-    #[trace]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_file_ops(
-        #[future]
-        #[notrace]
-        service: KubeService,
-        #[values(Agent::Ephemeral, Agent::Job)] agent: Agent,
-        #[values(FileOps::Python, FileOps::Go18, FileOps::Go19, FileOps::Rust)] ops: FileOps,
-    ) {
-        let service = service.await;
-        let _ = std::fs::create_dir(std::path::Path::new("/tmp/fs"));
-        let command = ops.command();
-
-        let mut args = vec!["--fs-mode", "write"];
-
-        if let Some(ephemeral_flag) = agent.flag() {
-            args.extend(ephemeral_flag);
-        }
-
-        let env = vec![("MIRRORD_FILE_READ_WRITE_PATTERN", "/tmp/**")];
-        let mut process = run(
-            command,
-            &service.target,
-            Some(&service.namespace),
-            Some(args),
-            Some(env),
-        )
-        .await;
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        ops.assert(process);
-    }
-
-    #[cfg(target_os = "macos")]
-    #[rstest]
-    #[trace]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_file_ops(
-        #[future]
-        #[notrace]
-        service: KubeService,
-        #[values(Agent::Job)] agent: Agent,
-    ) {
-        let service = service.await;
-        let _ = std::fs::create_dir(std::path::Path::new("/tmp/fs"));
-        let python_command = vec!["python3", "-B", "-m", "unittest", "-f", "python-e2e/ops.py"];
-        let args = vec!["--fs-mode", "read"];
-        let env = vec![("MIRRORD_FILE_READ_WRITE_PATTERN", "/tmp/fs/**")];
-
-        let mut process = run(
-            python_command,
-            &service.target,
-            Some(&service.namespace),
-            Some(args),
-            Some(env),
-        )
-        .await;
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_python_fileops_stderr();
-    }
-
-    #[rstest]
-    #[trace]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_file_ops_ro(
-        #[future]
-        #[notrace]
-        service: KubeService,
-    ) {
-        let service = service.await;
-        let _ = std::fs::create_dir(std::path::Path::new("/tmp/fs"));
-        let python_command = vec![
-            "python3",
-            "-B",
-            "-m",
-            "unittest",
-            "-f",
-            "python-e2e/files_ro.py",
-        ];
-
-        let mut process = run(
-            python_command,
-            &service.target,
-            Some(&service.namespace),
-            None,
-            None,
-        )
-        .await;
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_python_fileops_stderr();
-    }
-
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_remote_env_vars_exclude_works(#[future] service: KubeService) {
-        let service = service.await;
-        let node_command = vec![
-            "node",
-            "node-e2e/remote_env/test_remote_env_vars_exclude_works.mjs",
-        ];
-        let mirrord_args = vec!["-x", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(
-            node_command,
-            &service.target,
-            None,
-            Some(mirrord_args),
-            None,
-        )
-        .await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_remote_env_vars_include_works(#[future] service: KubeService) {
-        let service = service.await;
-        let node_command = vec![
-            "node",
-            "node-e2e/remote_env/test_remote_env_vars_include_works.mjs",
-        ];
-        let mirrord_args = vec!["-s", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(
-            node_command,
-            &service.target,
-            None,
-            Some(mirrord_args),
-            None,
-        )
-        .await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
 
     #[rstest]
     #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
@@ -314,137 +76,6 @@ mod tests {
             .unwrap();
 
         application.assert(&process);
-    }
-
-    #[rstest]
-    #[cfg(target_os = "linux")]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_remote_env_vars_works(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/env.sh"];
-        let mut process = run(bash_command, &service.target, None, None, None).await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[rstest]
-    #[cfg(target_os = "linux")]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_remote_env_vars_exclude_works(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/env.sh", "exclude"];
-        let mirrord_args = vec!["-x", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(
-            bash_command,
-            &service.target,
-            None,
-            Some(mirrord_args),
-            None,
-        )
-        .await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[rstest]
-    #[cfg(target_os = "linux")]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_remote_env_vars_include_works(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/env.sh", "include"];
-        let mirrord_args = vec!["-s", "MIRRORD_FAKE_VAR_FIRST"];
-        let mut process = run(
-            bash_command,
-            &service.target,
-            None,
-            Some(mirrord_args),
-            None,
-        )
-        .await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    // Currently fails due to Layer >> AddressConversion in ci for some reason
-
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_file_exists(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/file.sh", "exists"];
-        let mut process = run(bash_command, &service.target, None, None, None).await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    // currently there is an issue with piping across forks of processes so 'test_bash_file_read'
-    // and 'test_bash_file_write' cannot pass
-
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_file_read(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/file.sh", "read"];
-        let mut process = run(bash_command, &service.target, None, None, None).await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[ignore]
-    #[cfg(target_os = "linux")]
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    #[timeout(Duration::from_secs(240))]
-    pub async fn test_bash_file_write(#[future] service: KubeService) {
-        let service = service.await;
-        let bash_command = vec!["bash", "bash-e2e/file.sh", "write"];
-        let args = vec!["--rw"];
-        let mut process = run(bash_command, &service.target, None, Some(args), None).await;
-
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn test_go18_remote_env_vars_works(#[future] service: KubeService) {
-        let service = service.await;
-        let command = vec!["go-e2e-env/18"];
-        let mut process = run(command, &service.target, None, None, None).await;
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
-    }
-
-    #[rstest]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    pub async fn test_go19_remote_env_vars_works(#[future] service: KubeService) {
-        let service = service.await;
-        let command = vec!["go-e2e-env/19"];
-        let mut process = run(command, &service.target, None, None, None).await;
-        let res = process.child.wait().await.unwrap();
-        assert!(res.success());
-        process.assert_stderr();
     }
 
     // TODO: change outgoing TCP tests to use the same setup as in the outgoing UDP test so that
